% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapClustering.R
\name{addUmapTaxaColors}
\alias{addUmapTaxaColors}
\title{Add colors for taxa in UMAP plot}
\usage{
addUmapTaxaColors(plotDf = NULL, colorPalette = "Set2", 
    highlightTaxa = NULL)
}
\arguments{
\item{plotDf}{data for UMAP plot}

\item{colorPalette}{color palette. Default: "Set2"}

\item{highlightTaxa}{list of taxa to be highlighted}
}
\value{
A dataframe for UMAP plot with an additional column for the assigned
color to each taxon
}
\description{
Add colors for taxa in UMAP plot
}
\examples{
rawInput <- system.file(
   "extdata", "test.main.long", package = "PhyloProfile", mustWork = TRUE
)
longDf <- createLongMatrix(rawInput)
umapData <- prepareUmapData(longDf, "phylum")
data.umap <- umapClustering(umapData)
plotDf <- createUmapPlotData(data.umap, umapData)
PhyloProfile:::addUmapTaxaColors(plotDf, colorPalette = "Set2")
}
\seealso{
\code{\link{prepareUmapData}}, \code{\link{umapClustering}},
\code{\link{createUmapPlotData}}
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
