% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{calculatePValue}
\alias{calculatePValue}
\title{calculate the pValues}
\usage{
calculatePValue(varIn, varOut, significanceLevel)
}
\arguments{
\item{varIn}{list of values for the variable concerning the in-group}

\item{varOut}{list of values for the variable concerning the out-group}

\item{significanceLevel}{significant cutoff for statistical test}
}
\value{
return the pvalues
}
\description{
calculate the pValues
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
