% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{check_oma_id}
\alias{check_oma_id}
\title{check OMA IDs or Uniprot IDs as valid Input --------------------------------}
\usage{
check_oma_id(ids)
}
\arguments{
\item{ids}{list of ids needs to be checked}
}
\value{
list of invalid IDs (not readable for OMA)
}
\description{
check OMA IDs or Uniprot IDs as valid Input --------------------------------
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
