% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createColorScheme.R
\name{checkColorPalette}
\alias{checkColorPalette}
\title{Check if a color pallete has enough colors for a list of items}
\usage{
checkColorPalette(items, pallete = "Paired")
}
\arguments{
\item{items}{vector contains list of items}

\item{pallete}{name of color palette}
}
\value{
TRUE if color pallete has enough colors, otherwise FALSE
}
\description{
Check if a color pallete has enough colors for a list of items
}
\examples{
myItems <- rep("a",3)
checkColorPalette(myItems, "Set1")
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
