% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTaxonomyMatrix.R
\name{checkIdNewtaxa}
\alias{checkIdNewtaxa}
\title{Get valid NCBI taxonomy ID for an pseudo ID}
\usage{
checkIdNewtaxa(id = NULL, newTaxa = NULL)
}
\arguments{
\item{id}{pseudo NCBI taxonomy ID}

\item{newTaxa}{dataframe of the newtaxa.txt file that links pseudo IDs with
valid NCBI taxonomy IDs (default: /PhyloProfile/data/newTaxa.txt)}
}
\value{
A valid NCBI taxonomy ID
}
\description{
Get valid NCBI taxonomy ID for an pseudo ID
}
\examples{
id <- 9999999999
checkIdNewtaxa(id)
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
