% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMainInput.R
\name{checkInputValidity}
\alias{checkInputValidity}
\title{Check the validity of the main input file}
\usage{
checkInputValidity(filein)
}
\arguments{
\item{filein}{input file}
}
\value{
The format of the input file format, or type of error
}
\description{
Check if input file has one of the following format: orthoXML,
multiple FASTA, wide or long matrix, or a list of OMA IDs.
}
\examples{
filein <- system.file(
    "extdata", "test.main.wide", package = "PhyloProfile", mustWork = TRUE
)
checkInputValidity(filein)
}
\seealso{
\code{\link{checkOmaID}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
