% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{checkOmaID}
\alias{checkOmaID}
\title{Check OMA IDs}
\usage{
checkOmaID(ids)
}
\arguments{
\item{ids}{list of ids needs to be checked}
}
\value{
list of invalid IDs (not readable for OMA)
}
\description{
Check if input IDs are valid IDs for OMA Browser
(either OMA IDs or UniProt IDs)
}
\examples{
checkOmaID("HUMAN29398")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
