% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{compareMedianTaxonGroups}
\alias{compareMedianTaxonGroups}
\title{Compare the median values of a variable between 2 taxon groups}
\usage{
compareMedianTaxonGroups(data, inGroup, useCommonAncestor, variable)
}
\arguments{
\item{data}{input phylogenetic profile in long format (see ?mainLongRaw and 
?createLongMatrix)}

\item{inGroup}{ID list of in-group taxa (e.g. "ncbi1234")}

\item{useCommonAncestor}{TRUE/FALSE if using all taxa that share the same 
common ancestor with the pre-selected in-group as the in-group taxa. 
Default = TRUE.}

\item{variable}{name of the variable that need to be compared}
}
\value{
List of genes that have a difference in the variable's median scores
between the in-group and out-group taxa and their corresponding delta-median.
}
\description{
Given the phylogenetic profiles that contains up to 2 additional
variables besides the presence/absence information of the orthologous 
proteins. This function will compare the median scores of those variables 
between 2 different taxon groups (e.g. parasitic species vs non-parasitic 
species), which are defined as in-group and out-group. In-group is identified
by the user. Out-group contains all taxa in the input phylogenetic profiles
that are not part of the in-group.
}
\examples{
data("mainLongRaw", package="PhyloProfile")
data <- mainLongRaw
inGroup <- c("ncbi876142", "ncbi586133")
variable <- colnames(data)[4]
compareMedianTaxonGroups(data, inGroup, TRUE, variable)
}
\author{
Vinh Tran (tran@bio.uni-frankfurt.de)
}
