% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{create_domain_df}
\alias{create_domain_df}
\title{create domain annotation dataframe for ONE protein}
\usage{
create_domain_df(domain_id, ortho_id, length, domain_list)
}
\arguments{
\item{domain_id}{protein domain ID}

\item{ortho_id}{protein ID}

\item{length}{protein length}

\item{domain_list}{list of all domains and their positions for this protein}
}
\value{
domain annotation in a dataframe
}
\description{
create domain annotation dataframe for ONE protein
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
