% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{create_profile_from_oma}
\alias{create_profile_from_oma}
\title{create phylogenetic profile from full OMA dataframe}
\usage{
create_profile_from_oma(final_oma_df)
}
\arguments{
\item{final_oma_df}{raw OMA data for a list of input IDs
(data for each oma ID were obtained by get_data_for_one_ortholog() function)}
}
\value{
profile in long format
}
\description{
create phylogenetic profile from full OMA dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
