% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{createDomainDf}
\alias{createDomainDf}
\title{Create domain annotation dataframe for one OMA protein}
\usage{
createDomainDf(domainID, orthoID, length, domainList)
}
\arguments{
\item{domainID}{protein domain ID}

\item{orthoID}{protein ID}

\item{length}{protein length}

\item{domainList}{list of all domains and their positions for this protein}
}
\value{
domain annotation in a dataframe
}
\description{
Create domain annotation dataframe for one OMA protein
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
