% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMainInput.R
\name{createLongMatrix}
\alias{createLongMatrix}
\title{Create a long matrix format for all kinds of input file}
\usage{
createLongMatrix(inputFile)
}
\arguments{
\item{inputFile}{input file in orthoXML, multiple FASTA, wide or long
matrix format.}
}
\value{
A data frame that contains input data in long-format
}
\description{
Create a long matrix format for all kinds of input file
}
\examples{
inputFile <- system.file(
    "extdata", "test.main.wide", package = "phyloprofile", mustWork = TRUE
)
createLongMatrix(inputFile)
}
\seealso{
\code{\link{xmlParser}}, \code{\link{fastaParser}},
\code{\link{wideToLong}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
