% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{createProfileData}
\alias{createProfileData}
\title{Create data for plotting the phylogentic profiles}
\usage{
createProfileData(superTaxonData, refTaxon, percentCutoff,
    coorthologCutoffMax, var1Cutoff, var2Cutoff, var1Relation,
    var2Relation, groupByCat, catDt)
}
\arguments{
\item{superTaxonData}{a reduced dataframe contains info for all profiles in
the selected taxonomy rank.}

\item{refTaxon}{selected reference taxon}

\item{percentCutoff}{min and max cutoffs for percentage of species present
in a supertaxon}

\item{coorthologCutoffMax}{maximum number of co-orthologs allowed}

\item{var1Cutoff}{min and max cutoffs for var1}

\item{var2Cutoff}{min anc max cutoffs for var2}

\item{var1Relation}{relation of var1 ("protein" for protein-protein or
"species" for protein-species)}

\item{var2Relation}{relation of var2 ("protein" for protein-protein or
"species" for protein-species)}

\item{groupByCat}{group genes by their categories (TRUE or FALSE)}

\item{catDt}{dataframe contains gene categories
(optional, NULL if groupByCat = FALSE or no info provided)}
}
\value{
A dataframe ready for generating profile plot.
}
\description{
Create data for plotting the phylogentic profiles
}
\examples{
data("fullProcessedProfile", package="phyloprofile")
superTaxonData <- reduceProfile(fullProcessedProfile)
refTaxon <- "Mammalia"
percentCutoff <- c(0.0, 1.0)
coorthologCutoffMax <- 10
var1Cutoff <- c(0.75, 1.0)
var2Cutoff <- c(0.5, 1.0)
var1Relation <- "protein"
var2Relation <- "species"
groupByCat <- FALSE
catDt <- NULL
createProfileData(
    superTaxonData,
    refTaxon,
    percentCutoff,
    coorthologCutoffMax,
    var1Cutoff,
    var2Cutoff,
    var1Relation,
    var2Relation,
    groupByCat,
    catDt
)
}
\seealso{
\code{\link{parseInfoProfile}} and \code{\link{reduceProfile}}
for generating input dataframe, \code{\link{fullProcessedProfile}} for a
demo full processed profile dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
