% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{createProfileFromOma}
\alias{createProfileFromOma}
\title{Create phylogenetic profile from a raw OMA dataframe}
\usage{
createProfileFromOma(finalOmaDf)
}
\arguments{
\item{finalOmaDf}{raw OMA data for a list of input IDs}
}
\value{
phylogenetic profiles in long format
}
\description{
Create phylogenetic profile from a raw OMA dataframe
}
\examples{
omaData <- getDataForOneOma("HUMAN29397", "OG")
createProfileFromOma(omaData)
}
\seealso{
\code{\link{getDataForOneOma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
