% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhylotree.R
\name{createUnrootedTree}
\alias{createUnrootedTree}
\title{Create unrooted tree from a taxonomy matrix}
\usage{
createUnrootedTree(df)
}
\arguments{
\item{df}{data frame contains taxonomy matrix used for generating tree}
}
\value{
A unrooted taxonomy tree as an object of class "phylo".
}
\description{
Create unrooted tree from a taxonomy matrix
}
\examples{
data("ppTaxonomyMatrix", package = "PhyloProfile")
createUnrootedTree(ppTaxonomyMatrix)
}
\seealso{
\code{\link{taxa2dist}} for distance matrix generation from a
taxonomy matrix, \code{\link{getTaxonomyMatrix}} for getting taxonomy
matrix, \code{\link{ppTaxonomyMatrix}} for a demo taxonomy matrix data
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
