% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDistribution.R
\name{createVariableDistributionDataSubset}
\alias{createVariableDistributionDataSubset}
\title{Create data for additional variable distribution (for a subset data)}
\usage{
createVariableDistributionDataSubset(fullProfileData,
    distributionData, selectedGenes, selectedTaxa)
}
\arguments{
\item{fullProfileData}{dataframe contains the full processed profiles}

\item{distributionData}{dataframe contains the full distribution data}

\item{selectedGenes}{list of genes of interst}

\item{selectedTaxa}{list of taxa of interest}
}
\value{
A dataframe ready for analysing the distribution of the additional
variable(s) for a subset of genes and/or taxa.
}
\description{
Create data for additional variable distribution (for a subset data)
}
\examples{
data("fullProcessedProfile", package="phyloprofile")
data("mainLongRaw", package="phyloprofile")
distributionData <- createVariableDistributionData(
    mainLongRaw, 0, 1, 0.5, 1
)
selectedGenes <- "OG_1019"
selectedTaxa <- c("Mammalia", "Echinoidea", "Gunneridae", "Mucorales",
"Alphaproteobacteria")
createVariableDistributionDataSubset(
    fullProcessedProfile,
    distributionData,
    selectedGenes,
    selectedTaxa
)
}
\seealso{
\code{\link{parseInfoProfile}},
\code{\link{createVariableDistributionData}},
\code{\link{fullProcessedProfile}}, \code{\link{mainLongRaw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
