% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProfilePlot.R
\name{dataCustomizedPlot}
\alias{dataCustomizedPlot}
\title{Create data for customized profile plot}
\usage{
dataCustomizedPlot(dataHeat, selectedTaxa, selectedSeq)
}
\arguments{
\item{dataHeat}{a data frame contains processed profiles (see
?fullProcessedProfile, ?filterProfileData)}

\item{selectedTaxa}{selected subset of taxa}

\item{selectedSeq}{selected subset of genes}
}
\value{
A dataframe contains data for plotting the customized profile.
}
\description{
Create data for customized profile plot based on a selected
list of genes and/or taxa.
}
\examples{
data("fullProcessedProfile", package="PhyloProfile")
selectedTaxa <- c("Mammalia", "Echinoidea", "Gunneridae")
selectedSeq <- "all"
dataCustomizedPlot(fullProcessedProfile, selectedTaxa, selectedSeq)
}
\seealso{
\code{\link{filterProfileData}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
