% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProfilePlot.R
\name{dataMainPlot}
\alias{dataMainPlot}
\title{Create data for main profile plot}
\usage{
dataMainPlot(dataHeat)
}
\arguments{
\item{dataHeat}{a data frame contains processed profiles (see
?fullProcessedProfile, ?filterProfileData)}
}
\value{
A dataframe contains data for plotting the phylogenetic profile.
}
\description{
Create data for main profile plot
}
\examples{
data("fullProcessedProfile", package="PhyloProfile")
dataMainPlot(fullProcessedProfile)
}
\seealso{
\code{\link{filterProfileData}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
