% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{distributionTest}
\alias{distributionTest}
\title{Compare the distribution of 2 numeric vectors}
\usage{
distributionTest(varIn, varOut, significanceLevel)
}
\arguments{
\item{varIn}{first numeric vector}

\item{varOut}{second numeric vector}

\item{significanceLevel}{significant cutoff of the Kolmogorov-Smirnov test. 
Default = 0.05.}
}
\value{
p-value of the comparison test.
}
\description{
This function tests the difference between the distributions of 
two input numeric samples using the statistical tess. First the 
Kolmogorov-Smirnov is used to check if 2 samples have the same distribution.
If yes, Wilcoxon-Mann-Whitney will be used to compare the distribution 
difference.
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
