% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGeneAge.R
\name{estimateGeneAge}
\alias{estimateGeneAge}
\title{Calculate the phylogenetic gene age from profiles}
\usage{
estimateGeneAge(processedProfileData, rankName, refTaxon,
    var1Cutoff, var2Cutoff, percentCutoff)
}
\arguments{
\item{processedProfileData}{dataframe contains the full processed
phylogenetic profiles}

\item{rankName}{taxonomy rank (e.g. "species", "genus", "family")}

\item{refTaxon}{reference taxon (e.g. "Homo sapiens", "Homo", "Hominidae")}

\item{var1Cutoff}{cutoff for var1}

\item{var2Cutoff}{cutoff for var2}

\item{percentCutoff}{cutoff for percentage of species present in each
supertaxon}
}
\value{
A dataframe contains estimated gene ages
}
\description{
Calculate the phylogenetic gene age from profiles
}
\examples{
data("fullProcessedProfile", package="PhyloProfile")
rankName <- "class"
refTaxon <- "Mammalia"
processedProfileData <- fullProcessedProfile
var1Cutoff <- c(0, 1)
var2Cutoff <- c(0, 1)
percentCutoff <- c(0, 1)
estimateGeneAge(
    processedProfileData,
    rankName,
    refTaxon,
    var1Cutoff, var2Cutoff, percentCutoff
)
}
\seealso{
\code{\link{parseInfoProfile}} for creating a full processed
profile dataframe; \code{\link{getNameList}} and
\code{\link{getTaxonomyMatrix}} for getting taxonomy info,
\code{\link{fullProcessedProfile}} for a demo input dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
