% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapClustering.R
\name{fallbackUmap}
\alias{fallbackUmap}
\title{Fallback for UMAP in case of insufficient samples}
\usage{
fallbackUmap(umapDt, randomSeed, dim)
}
\arguments{
\item{umapDt}{data matrix for UMAP}

\item{randomSeed}{random seed. Default: 123}

\item{dim}{dimension, either 2 for 2D (default) or 3 for 3D}
}
\value{
A table contains coordinates UMAP reduction
}
\description{
Fallback for UMAP in case of insufficient samples
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
