% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMainInput.R
\name{fastaParser}
\alias{fastaParser}
\title{Parse multi-fasta input file}
\usage{
fastaParser(inputFile)
}
\arguments{
\item{inputFile}{input multiple fasta file}
}
\value{
A data frame containing input data in long-format
}
\description{
Parse multi-fasta input file
}
\examples{
\dontrun{
inputFile <- system.file(
    "extdata", "test.main.fasta", package = "PhyloProfile", mustWork = TRUE
)
fastaParser(inputFile)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
