% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{filterProfileData}
\alias{filterProfileData}
\title{Filter phylogentic profiles}
\usage{
filterProfileData(superTaxonData, refTaxon, percentCutoff,
    coorthologCutoffMax, var1Cutoff, var2Cutoff, var1Relation,
    var2Relation, groupByCat, catDt)
}
\arguments{
\item{superTaxonData}{a reduced dataframe contains info for all phylogenetic
profiles in the selected taxonomy rank.}

\item{refTaxon}{selected reference taxon}

\item{percentCutoff}{min and max cutoffs for percentage of species present
in a supertaxon}

\item{coorthologCutoffMax}{maximum number of co-orthologs allowed}

\item{var1Cutoff}{min and max cutoffs for var1}

\item{var2Cutoff}{min anc max cutoffs for var2}

\item{var1Relation}{relation of var1 ("protein" for protein-protein or
"species" for protein-species)}

\item{var2Relation}{relation of var2 ("protein" for protein-protein or
"species" for protein-species)}

\item{groupByCat}{group genes by their categories (TRUE or FALSE)}

\item{catDt}{dataframe contains gene categories
(optional, NULL if groupByCat = FALSE or no info provided)}
}
\value{
A filtered dataframe for generating profile plot including seed gene 
IDs (or orthologous group IDs), their ortholog IDs and the corresponding 
(super)taxa, (super)taxon IDs, number of co-orthologs in each (super)taxon, 
values for two additional variables var1, var2, % of species present in each 
supertaxon, and the categories of seed genes (or ortholog groups).
}
\description{
Create a filtered data needed for plotting or clustering 
phylogenetic profiles. NOTE: this function require some intermediate steps 
using the results from other functions. If you would like to get a full 
processed data from the raw input, please use the function 
fromInputToProfile() instead!
}
\examples{
# NOTE: this function require some intermediate steps using the results from
# other functions. If you would like to get a full processed data from the
# raw input, please use the function fromInputToProfile() instead!
data("fullProcessedProfile", package="PhyloProfile")
superTaxonData <- reduceProfile(fullProcessedProfile)
refTaxon <- "Mammalia"
percentCutoff <- c(0.0, 1.0)
coorthologCutoffMax <- 10
var1Cutoff <- c(0.75, 1.0)
var2Cutoff <- c(0.5, 1.0)
var1Relation <- "protein"
var2Relation <- "species"
groupByCat <- FALSE
catDt <- NULL
filterProfileData(
    superTaxonData,
    refTaxon,
    percentCutoff,
    coorthologCutoffMax,
    var1Cutoff,
    var2Cutoff,
    var1Relation,
    var2Relation,
    groupByCat,
    catDt
)
}
\seealso{
\code{\link{parseInfoProfile}} and \code{\link{reduceProfile}}
for generating input dataframe, \code{\link{fullProcessedProfile}} for a
demo full processed profile dataframe, \code{\link{fromInputToProfile}} for
generating fully processed data from raw input.
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
