% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullProcessedProfile}
\alias{fullProcessedProfile}
\title{An example of a fully processed phylogenetic profile.}
\format{
A data frame with 168 rows and 17 variables:
\itemize{
    \item{supertaxon}{ Supertaxon name together with its ordered index, e.g.
    "1001_Mammalia"}
    \item{geneID}{ Seed or ortholog group ID, e.g. "100136at6656"}
    \item{ncbiID}{ Taxon ID, e.g. "ncbi10116"}
    \item{orthoID}{ Ortholog ID, e.g. "100136at6656|HUMAN@9606@1|Q9UNQ2|1"}
    \item{var1}{ First additional variable}
    \item{var2}{ Second additional variable}
    \item{paralog}{ Number of co-orthologs in the current taxon}
    \item{abbrName}{ NCBI ID of the ortholog, e.g. "ncbi9606"}
    \item{taxonID}{ Taxon ID of the ortholog, in this case: "0"}
    \item{fullName}{ Full taxon name of the ortholog, e.g. "Homo sapiens"}
    \item{supertaxonID}{ Supertaxon ID (only different than ncbiID in case
    working with higher taxonomy rank than input's). e.g. "40674"}
    \item{rank}{ Rank of the supertaxon, e.g. "class"}
    \item{category}{ "cat}
    \item{presSpec}{ The percentage of species presenting in each supertaxon}
    \item{mVar1}{ Value of the 1. variable after grouping into supertaxon}
    \item{mVar2}{ Value of the 2. variable after grouping into supertaxon}
    \item{numberSpec}{ Total number of species in each supertaxon}
}
}
\usage{
data(fullProcessedProfile)
}
\description{
An example of a fully processed phylogenetic profile.
}
