% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullProcessedProfileLarge}
\alias{fullProcessedProfileLarge}
\title{An example of a large processed phylogenetic profile.}
\format{A data frame with 80 rows and 9 variables:
\itemize{
    \item{geneID}{ Seed or ortholog group ID, e.g. "OG_1017"}
    \item{supertaxon}{ Supertaxon name together with its ordered index, e.g.
    "1001_Mammalia"}
    \item{supertaxonID}{ Supertaxon ID (only different than ncbiID in case
    working with higher taxonomy rank than input's). e.g. "40674"}
    \item{var1}{ First additional variable}
    \item{presSpec}{ The percentage of species presenting in each supertaxon}
    \item{category}{ "cat}
    \item{orthoID}{ Ortholog ID, e.g. "M.musculus@10090@112934"}
    \item{var2}{ Second additional variable}
    \item{paralog}{ Number of co-orthologs in the current taxon}
}}
\usage{
fullProcessedProfileLarge
}
\description{
An example of a large processed phylogenetic profile.
}
\keyword{datasets}
