% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_all_domains_oma}
\alias{get_all_domains_oma}
\title{create domain annotation dataframe for complete OMA data}
\usage{
get_all_domains_oma(final_oma_df)
}
\arguments{
\item{final_oma_df}{raw OMA data for a list of input IDs
(data for each oma ID were obtained by get_data_for_one_ortholog() function)}
}
\value{
domain annotation in a dataframe to input into phyloprofile
}
\description{
create domain annotation dataframe for complete OMA data
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
