% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_all_fasta_oma}
\alias{get_all_fasta_oma}
\title{get all fasta sequences for complete OMA data}
\usage{
get_all_fasta_oma(final_oma_df)
}
\arguments{
\item{final_oma_df}{obtained OMA data for a list of input IDs}
}
\value{
dataframe contains all fasta sequences
}
\description{
get all fasta sequences for complete OMA data
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
