% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_data_for_one_oma}
\alias{get_data_for_one_oma}
\title{get taxonomy ID, sequence and annotation for an OMA sequence ----------------}
\usage{
get_data_for_one_oma(seed_id, ortho_type)
}
\arguments{
\item{seed_id}{seed ID in OMA or UniProt format}

\item{ortho_type}{type of OMA orthologs}
}
\value{
data frame contains above info for ONE OMA id
}
\description{
get taxonomy ID, sequence and annotation for an OMA sequence ----------------
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
