% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_data_for_one_ortholog}
\alias{get_data_for_one_ortholog}
\title{get taxonomy ID, sequence and annotation for an OMA sequence ----------------}
\usage{
get_data_for_one_ortholog(id)
}
\arguments{
\item{id}{oma ID of a ortholog}
}
\value{
data frame contains above info for ONE OMA id
}
\description{
get taxonomy ID, sequence and annotation for an OMA sequence ----------------
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
