% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_domain_from_url}
\alias{get_domain_from_url}
\title{get domain annotation for a domain URL of an OMA sequence -------------------}
\usage{
get_domain_from_url(domainURL)
}
\arguments{
\item{domainURL}{URL address for domain annotation of ONE OMA id}
}
\value{
data frame contains feature names, start and end positions
}
\description{
get domain annotation for a domain URL of an OMA sequence -------------------
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
