% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxonomy_matrix.R
\name{get_ids_rank}
\alias{get_ids_rank}
\title{Get taxonomy IDs for a list of taxa}
\usage{
get_ids_rank(input_taxa, currentNCBIinfo)
}
\arguments{
\item{input_taxa}{input taxon list}

\item{currentNCBIinfo}{table of the current taxonomy data (taxonNamesFull.txt)}
}
\value{
list of 3 dataframes: idList, rankList and reducedInfoList
}
\description{
Get taxonomy IDs for a list of taxa
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
