% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_oma_members}
\alias{get_oma_members}
\title{get the members for a OMA or Uniprot id ------------------------------------}
\usage{
get_oma_members(id, ortho_type)
}
\arguments{
\item{id}{oma ID}

\item{ortho_type}{type of OMA orthologs}
}
\value{
list of ortholog members
}
\description{
get the members for a OMA or Uniprot id ------------------------------------
}
\author{
Carla Mölbert {carla.moelbert@gmx.de}
}
