% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_selected_fasta_oma}
\alias{get_selected_fasta_oma}
\title{get selected fasta sequences from the complete OMA data}
\usage{
get_selected_fasta_oma(final_oma_df, seq_id)
}
\arguments{
\item{final_oma_df}{obtained OMA data for a list of input IDs}

\item{seq_id}{sequence need to be returned}
}
\value{
required sequence in fasta format
}
\description{
get selected fasta sequences from the complete OMA data
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
