% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{getAllDomainsOma}
\alias{getAllDomainsOma}
\title{Create domain annotation dataframe from a raw OMA dataframe}
\usage{
getAllDomainsOma(finalOmaDf)
}
\arguments{
\item{finalOmaDf}{raw OMA data for a list of input IDs}
}
\value{
domain annotation in a dataframe to input into PhyloProfile, which
contains seed ID, ortholog ID, ortholog length, annotated feature, start
and end position of that feature.
}
\description{
Create domain annotation dataframe from a raw OMA dataframe
}
\examples{
omaData <- getDataForOneOma("HUMAN29397", "OG")
getAllDomainsOma(omaData)
}
\seealso{
\code{\link{getDataForOneOma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
