% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getCommonAncestor}
\alias{getCommonAncestor}
\title{Get common ancestor for a selected taxa}
\usage{
getCommonAncestor(inGroup, rank, nameList, taxonomyMatrix,
    ncbiIDList)
}
\arguments{
\item{inGroup}{list of selected taxa}

\item{rank}{taxonomy rank of input taxa (e.g. "species", "family", ...)}

\item{nameList}{list contains "ncbiID", "fullName", "rank", "parentID" of all
pre-calculated taxa (from /PhyloProfile/data/taxonNamesReduced.txt)}

\item{taxonomyMatrix}{pre-calculated taxonomy matrix (see 
/PhyloProfile/data/taxonomyMatrix.txt)}

\item{ncbiIDList}{list of input taxon IDs}
}
\value{
common anchestor
}
\description{
Get common ancestor for a selected taxa
}
\examples{
inGroup <- c("Aeropyrum pernix", "Agrobacterium fabrum")
rank <- "species"
nameListFile <- system.file(
    "extdata", "data/taxonNamesFull.txt",
    package = "PhyloProfile", mustWork = TRUE
)
nameList <- data.table::fread(nameListFile)
taxonomyMatrix <- getTaxonomyMatrix(FALSE, NULL)
ncbiIDList <- c(56636, 1176649)
getCommonAncestor(
    inGroup,
    rank,
    nameList,
    taxonomyMatrix,
    ncbiIDList
)
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
