% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfile.R
\name{getDataClustering}
\alias{getDataClustering}
\title{Get data for calculating the distance matrix}
\usage{
getDataClustering(data, profileType, var1AggregateBy,
    var2AggregateBy)
}
\arguments{
\item{data}{processed profile data}

\item{profileType}{type of data used for calculating the distance matrix.
Either "binary" (consider only the presence/absence status of orthlogs), or
"var1"/"var2" for taking values of the additional variables into account.}

\item{var1AggregateBy}{aggregate method for VAR1 (min, max, mean or median)}

\item{var2AggregateBy}{aggregate method for VAR2 (min, max, mean or median)}
}
\value{
A wide dataframe contains values for calculating distance matrix.
}
\description{
Get data for calculating the distance matrix
}
\note{
Documented by Vinh Tran (tran@bio.uni-frankfurt.de)
}
\examples{
data("fullProcessedProfile", package="PhyloProfile")
data <- fullProcessedProfile
profileType <- "binary"
var1AggregateBy <- "max"
var2AggregateBy <- "mean"
getDataClustering(data, profileType, var1AggregateBy, var2AggregateBy)
}
\seealso{
\code{\link{fromInputToProfile}}
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
