% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{getDataForOneOma}
\alias{getDataForOneOma}
\title{Get OMA info for a query protein and its orthologs}
\usage{
getDataForOneOma(seedID, orthoType)
}
\arguments{
\item{seedID}{protein query ID in OMA or UniProt format}

\item{orthoType}{type of OMA orthologs}
}
\value{
data frame contains info for all sequences of that OMA group (or HOG)
}
\description{
Get taxonomy IDs, sequences and annotations for an OMA
orthologous group (or OMA HOG).
}
\examples{
getDataForOneOma("HUMAN29397", "OG")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
