% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDomainInput.R
\name{getDomainFile}
\alias{getDomainFile}
\title{Get domain file from a single (concatenate) file}
\usage{
getDomainFile(inputFile)
}
\arguments{
\item{inputFile}{concatenate domain file}
}
\value{
Domain file and its complete directory path
}
\description{
Get domain file from a single (concatenate) file
}
\examples{
\dontrun{
inputFile <- system.file(
    "extdata", "domainFiles/OG_1009.domains",
    package = "phyloprofile", mustWork = TRUE
)
getDomainFile("lca-micors")
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
