% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDomainInput.R
\name{getDomainOnline}
\alias{getDomainOnline}
\title{Get domain file(s) for online data set}
\usage{
getDomainOnline(demoData)
}
\arguments{
\item{demoData}{demo data name (either lca-micros or ampk-tor)}
}
\value{
Domain file and its complete directory path
}
\description{
Get domain file(s) for online data set
}
\examples{
\dontrun{
getDomainOnline("lca-micors")
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
