% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastaSeqs.R
\name{getFastaSeqs}
\alias{getFastaSeqs}
\title{Get fasta sequences}
\usage{
getFastaSeqs(dataIn, filein, demoData, inputTypeFasta,
    concatFasta, path, dirFormat, fileExt, idFormat)
}
\arguments{
\item{dataIn}{a dataframe of the input phylogenetic profiles, that contains
at least 3 columns: geneIDs, orthoIDs and ncbiIDs}

\item{filein}{main input file (for checking input type)}

\item{demoData}{name of demo data set (either "ampk-tor", "lca-micros",
or "none")}

\item{inputTypeFasta}{source of fasta sequences ("Concatenated fasta file"
or "Fasta folder")}

\item{concatFasta}{input concatenated fasta file}

\item{path}{path to fasta folder}

\item{dirFormat}{directory format (either "path/speciesID.fa*" or
"path/speciesID/speciesID.fa*")}

\item{fileExt}{fasta file extension ("fa", "fasta", "fas" or "txt")}

\item{idFormat}{fasta header format (">speciesID:seqID",
">speciesID@seqID", ">speciesID|seqID" or only "seqID")}
}
\value{
A dataframe contains fasta sequences
}
\description{
Get fasta sequences
}
\examples{
data("mainLongRaw", package="phyloprofile")
dataIn <- mainLongRaw
filein <- system.file(
    "extdata", "test.main.long", package = "phyloprofile", mustWork = TRUE
)
demoData <- "none"
inputTypeFasta <- "Concatenated fasta file"
concatFasta <- system.file(
    "extdata", "fastaFiles/concatenatedFile.fa",
    package = "phyloprofile", mustWork = TRUE
)
# the following parameters are conly required if upload a fasta folder!
path <- NULL
dirFormat <- NULL
fileExt <- NULL
idFormat <- NULL
# get fasta sequences
getFastaSeqs(
    dataIn, filein, demoData,
    inputTypeFasta,
    concatFasta,
    path,
    dirFormat,
    fileExt,
    idFormat
)
}
\seealso{
\code{\link{checkInputValidity}}, \code{\link{strReverse}},
\code{\link{mainLongRaw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
