% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getFeatures}
\alias{getFeatures}
\title{get the list with all the features in the gene}
\usage{
getFeatures(selectedGene, domains)
}
\arguments{
\item{selectedGene}{gene to get the feartures for}

\item{domains}{contains "seedID", "orthoID", "feature", "start",   "end"}
}
\value{
dataframe for the specific gene containing "seedID",  "orthoID",
"feature", "start",   "end"
}
\description{
get the list with all the features in the gene
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
