% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTaxonomyMatrix.R
\name{getIDsRank}
\alias{getIDsRank}
\title{Get taxonomy info for a list of taxa}
\usage{
getIDsRank(inputTaxa, currentNCBIinfo)
}
\arguments{
\item{inputTaxa}{input taxon list}

\item{currentNCBIinfo}{table/dataframe of the current taxonomy data
(taxonNamesFull.txt)}
}
\value{
A list of 3 dataframes: idList, rankList and reducedInfoList
}
\description{
Get NCBI taxonomy IDs, ranks and names for an input
taxon list
}
\examples{
inputTaxa <- c("272557", "176299")
ncbiFilein <- system.file(
    "extdata", "data/taxonNamesFull.txt",
    package = "PhyloProfile", mustWork = TRUE
)
currentNCBIinfo <- as.data.frame(data.table::fread(ncbiFilein))
getIDsRank(inputTaxa, currentNCBIinfo)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
