% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{getInputTaxaID}
\alias{getInputTaxaID}
\title{Get ID list of input taxa from the main input}
\usage{
getInputTaxaID(rawProfile)
}
\arguments{
\item{rawProfile}{long dataframe of input profile}
}
\value{
List of all input taxon IDs
}
\description{
Get ID list of input taxa from the main input
}
\examples{
data("mainLongRaw", package="phyloprofile")
getInputTaxaID(mainLongRaw)
}
\seealso{
\code{\link{createLongMatrix}}, \code{\link{mainLongRaw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
