% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{getOmaMembers}
\alias{getOmaMembers}
\title{Get OMA members}
\usage{
getOmaMembers(id = NULL, orthoType = "OG")
}
\arguments{
\item{id}{ID of the seed protein (OMA or UniProt ID)}

\item{orthoType}{type of OMA orthologs: either "HOG", "OG"
(orthologous group) or "PAIR" (orthologous pair - CURRENTLY NOT WORKING).
Default = "OG".}
}
\value{
List of OMA orthologs for an input seed protein.
}
\description{
Get OMA ortholog group, OMA HOG or OMA pair's members for a
seed protein from OMA Browser.
}
\examples{
print("Uncomment the following line to run the function")
# getOmaMembers("HUMAN29397", "OG")
}
\author{
Carla Mölbert {carla.moelbert@gmx.de}
}
