% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getPrefixFeatures}
\alias{getPrefixFeatures}
\title{Get the database for each feature in a specific gene}
\usage{
getPrefixFeatures(data)
}
\arguments{
\item{data}{contains "seedID", "orthoID", "feature", "start", "end"}
}
\value{
list of prefixes for the features
}
\description{
Get the database for each feature in a specific gene
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
