% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getPValues}
\alias{getPValues}
\title{Decide if the gene is significant}
\usage{
getPValues(inGroup, outGroup, variable, gene, parameters)
}
\arguments{
\item{inGroup}{contains "supertaxon", "geneID", "ncbiID", "orthoID",
"var1", "var2", "paralog", "abbrName", "taxonID", "fullname",
"supertaxonID", "rank", "category", "presSpec", "mVar1", "mVar2"}

\item{outGroup}{as in-group but with information containing the out-group}

\item{variable}{variable(s) to claculate the plots for}

\item{gene}{gene to calculate the p-values for}

\item{parameters}{contains "showPValue","highlightSignificant",
"significance", "var1ID", "var2ID", "xSizeGC", "ySizeGC",
"interestingFeatures", "angleGC", "legendGC", "legendSizeGC"}
}
\value{
return the pvalues
}
\description{
Decide if the gene is significant
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
