% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createColorScheme.R
\name{getQualColForVector}
\alias{getQualColForVector}
\title{Get color for a list of items}
\usage{
getQualColForVector(x = NULL)
}
\arguments{
\item{x}{input list}
}
\value{
list of colors for each element (same elements will have the same
color)
}
\description{
Get color for a list of items
}
\examples{
items <- c("a", "b", "c")
getQualColForVector(items)
}
\seealso{
\code{\link{qualitativeColours}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
