% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{getSelectedFastaOma}
\alias{getSelectedFastaOma}
\title{Get selected fasta sequences from a raw OMA dataframe}
\usage{
getSelectedFastaOma(finalOmaDf, seqID)
}
\arguments{
\item{finalOmaDf}{raw OMA data for a list of input IDs}

\item{seqID}{sequence need to be returned}
}
\value{
required sequence in fasta format
}
\description{
Get selected fasta sequences from a raw OMA dataframe
}
\examples{
omaData <- getDataForOneOma("HUMAN29397", "OG")
getSelectedFastaOma(omaData, "HUMAN29397")
}
\seealso{
\code{\link{getDataForOneOma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
