% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getSelectedSubset}
\alias{getSelectedSubset}
\title{Get the subset depending on the choosen rank}
\usage{
getSelectedSubset(rank, inGroup, nameList, taxaList)
}
\arguments{
\item{rank}{selected rank}

\item{inGroup}{list of taxa}

\item{nameList}{contains "ncbiID", "fullName", "rank", "parentID"}

\item{taxaList}{contains "abbrName, "ncbiID", fullName", "strain", "genus"}
}
\value{
list of prefixes for the features
}
\description{
Get the subset depending on the choosen rank
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
