% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getSignificantGenes}
\alias{getSignificantGenes}
\title{Get candidate genes of the taxonomy group comparison}
\usage{
getSignificantGenes(inGroup, selectedGenesList, rank, var,
    useCommonAncestor, referenceTaxon, parameters, ncbiIDList,
    dataFull, rightFormatFeatures, domains)
}
\arguments{
\item{inGroup}{list of taxa (full or subset) used as in-group}

\item{selectedGenesList}{list of genes that need to be compared}

\item{rank}{working taxonamy rank (e.g. "species", "family", "phylum", etc.)}

\item{var}{variable(s) used for comparison ("var1", "var2" or "both")}

\item{useCommonAncestor}{TRUE/FALSE if the common anchestor of all taxa in
the submitted inGroup should be used as in-group}

\item{referenceTaxon}{name of reference taxon in the in-group (e.g.
"Homo sapiens")}

\item{parameters}{options for plotting the distribution plots, containing
"showPValue" (TRUE/FALSE show/hide test's P-value on the distribution plot),
"highlightSignificant" (highlight distribution plot, which shows the
significant difference), "significance" (significant level, default = 0.05),
"var1ID", "var2ID", "xSizeGC", "ySizeGC", "interestingFeatures" (subset of
annotation features that need to be plotted), "angleGC", "legendGC",
"legendSizeGC"}

\item{ncbiIDList}{list of all ncbi IDs of input taxa}

\item{dataFull}{full processed main data (see ?fullProcessedProfile and
?parseInfoProfile)}

\item{rightFormatFeatures}{TRUE/FALSE if the features have the right format}

\item{domains}{dataframe contains the input protein domain}
}
\value{
Dataframe contains the genes that show significant difference in the
distributions of selected variables between the in-group and out-group.
}
\description{
Compare the distributions of additional variables (var1, var2)
between two taxonomy groups (in- and out-group) and return a list of genes
that shows the significant difference between those two groups. The taxonomy
in-group can be manually selected or generated by taking the last common
ancestor of the input taxa.
}
\examples{
data("mainLongRaw", package="PhyloProfile")
data("fullProcessedProfile", package="PhyloProfile")
inGroup <- c("Aeropyrum pernix", "Agrobacterium fabrum")
selectedGenesList <- "OG_1017"
rank <- "species"
var <- ""
useCommonAncestor <- FALSE
referenceTaxon <- "Homo sapiens"
ncbiIDList <- getInputTaxaID(mainLongRaw)
dataFull <- fullProcessedProfile
rightFormatFeatures <- TRUE
domainFile <- system.file(
    "extdata", "domainFiles/OG_1017.domains",
    package = "PhyloProfile", mustWork = TRUE
)
domains <- parseDomainInput("OG_1017", domainFile, "file")
parameters <- list(
    "showPValue" = TRUE,
    "highlightSignificant" = FALSE,
    "significance" = 0.05,
    "var1ID" = "1st variable",
    "var2ID" = "2nd variable",
    "xSizeGC" = 9,
    "ySizeGC" = 9,
    "interestingFeatures" = NULL,
    "angleGC" = 90,
    "legendGC" = "none",
    "legendSizeGC" = 9,
    "pValuesSize" = 9
)
getSignificantGenes(
    inGroup,
    selectedGenesList,
    rank,
    var,
    useCommonAncestor,
    referenceTaxon,
    parameters,
    ncbiIDList,
    dataFull,
    rightFormatFeatures,
    domains
)
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
