% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{getTaxonomyMatrix}
\alias{getTaxonomyMatrix}
\title{Get taxonomy matrix}
\usage{
getTaxonomyMatrix(subsetTaxaCheck, taxonIDs)
}
\arguments{
\item{subsetTaxaCheck}{subset taxonomy matrix based on input taxon IDs
(TRUE/FALSE)}

\item{taxonIDs}{list of input taxon IDs}
}
\value{
Data frame contains the (subset of) taxonomy matrix
}
\description{
Get full taxonomy matrix from "data/taxonomyMatrix.txt" or
only a subset of matrix based on an input taxon list
}
\examples{
# get full pre-installed taxonomy matrix
getTaxonomyMatrix(FALSE, NULL)
# get taxonomy matrix for a list of taxon IDs
taxonIDs <- c("ncbi10020", "ncbi10090")
getTaxonomyMatrix(TRUE, taxonIDs)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
