% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getTaxonomyRanks}
\alias{getTaxonomyRanks}
\title{Create a list with all main taxanomy ranks}
\usage{
getTaxonomyRanks()
}
\value{
A list of all main ranks (from strain to superkingdom)
}
\description{
Create a list with all main taxanomy ranks
}
\examples{
getTaxonomyRanks()
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
