% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapClustering.R
\name{groupLabelUmapData}
\alias{groupLabelUmapData}
\title{Reduce the number of labels for UMAP plot based on the gene/taxon frequency}
\usage{
groupLabelUmapData(data4umap = NULL, freqCutoff = 0)
}
\arguments{
\item{data4umap}{data for UMAP clustering (output from prepareUmapData)}

\item{freqCutoff}{gene/taxon frequency cutoff}
}
\value{
A dataframe similar to input data4umap, but with modified Label 
column, where less frequent labels are grouped together as "Other"
}
\description{
Reduce the number of labels for UMAP plot based on the gene/taxon frequency
}
\examples{
rawInput <- system.file(
   "extdata", "test.main.long", package = "PhyloProfile", mustWork = TRUE
)
longDf <- createLongMatrix(rawInput)
data4umap <- prepareUmapData(longDf, "phylum")
groupLabelUmapData(data4umap, freqCutoff = 3)
}
\seealso{
\code{\link{prepareUmapData}}
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
