% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainLongRaw}
\alias{mainLongRaw}
\title{An example of a raw long input file.}
\format{A data frame with 20 rows and 5 variables:
\itemize{
    \item{geneID}{ Seed or ortholog group ID, e.g. "OG_1017"}
    \item{ncbiID}{ Taxon ID, e.g. "ncbi176299"}
    \item{orthoID}{ Ortholog ID, e.g. "A.fabrum@176299@1582"}
    \item{FAS}{ First additional variable}
    \item{traceability}{ Second additional variable}
}}
\usage{
mainLongRaw
}
\description{
An example of a raw long input file.
}
\keyword{datasets}
