% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTaxonomyMatrix.R
\name{mainTaxonomyRank}
\alias{mainTaxonomyRank}
\title{Get all NCBI taxonomy rank names}
\usage{
mainTaxonomyRank()
}
\value{
A list of all available NCBI taxonomy rank names.
}
\description{
Get all NCBI taxonomy rank names
}
\examples{
mainTaxonomyRank()
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
