% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDomainInput.R
\name{parseDomainInput}
\alias{parseDomainInput}
\title{Parse domain input file}
\usage{
parseDomainInput(seed, inputFile, type)
}
\arguments{
\item{seed}{seed ID(s)}

\item{inputFile}{name of input data (demo data, file name or path to folder)}

\item{type}{type of data (demo, file or folder)}
}
\value{
A dataframe containing protein domain info
}
\description{
Parse domain input file
}
\examples{
seed <- "OG_1009"
inputFile <- system.file(
    "extdata", "domainFiles/OG_1009.domains",
    package = "PhyloProfile", mustWork = TRUE
)
type <- "file"
parseDomainInput(seed, inputFile, type)
}
\seealso{
\code{\link{getDomainOnline}}, \code{\link{getDomainFile}},
\code{\link{getDomainFolder}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
