% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapClustering.R
\name{performPCA}
\alias{performPCA}
\title{Helper function to perform PCA}
\usage{
performPCA(pcaDt)
}
\arguments{
\item{pcaDt}{data matrix for PCA}
}
\value{
A table contains coordinates of the first 3 PCs
}
\description{
Helper function to perform PCA
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
