% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapClustering.R
\name{performUmap}
\alias{performUmap}
\title{Helper function to handle UMAP logic}
\usage{
performUmap(umapDt, randomSeed = 123, dim = 2)
}
\arguments{
\item{umapDt}{data matrix for UMAP}

\item{randomSeed}{random seed. Default: 123}

\item{dim}{dimension, either 2 for 2D (default) or 3 for 3D}
}
\value{
A table contains coordinates UMAP reduction
}
\description{
Helper function to handle UMAP logic
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
