% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppTaxonomyMatrix}
\alias{ppTaxonomyMatrix}
\title{An example of a taxonomy matrix.}
\format{
A data frame with 10 rows and 162 variables:
\itemize{
    \item{abbrName}{ e.g. "ncbi10090"}
    \item{ncbiID}{ e.g. "10090"}
    \item{fullName}{ e.g. "Mus musculus"}
    \item{strain}{ e.g. "10090"}
    ...
}
}
\usage{
ppTaxonomyMatrix
}
\description{
An example of a taxonomy matrix.
}
\keyword{datasets}
