% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppTree}
\alias{ppTree}
\title{An example of a taxonomy tree in newick format}
\format{
Dataframe
}
\usage{
data(ppTree)
}
\value{
A data frame with only one entry
\itemize{
    \item V1 tree in newick format
}
}
\description{
An example of a taxonomy tree in newick format
}
