% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProfilePlot.R
\name{processOrthoID}
\alias{processOrthoID}
\title{Process ortholog IDs}
\usage{
processOrthoID(dataHeat = NULL)
}
\arguments{
\item{dataHeat}{a data frame contains processed profiles (see
?fullProcessedProfile, ?filterProfileData)}
}
\value{
the same dataframe as input, but the ortholog IDs are changed into
<taxID:orthoID>. New column {orthoFreq} specifies if the ortholog IDs are
single or duplicated
}
\description{
Process ortholog IDs to identify duplicated IDs
}
\examples{
?processOrthoID
\dontrun{
data("finalProcessedProfile", package="PhyloProfile")
processOrthoID(finalProcessedProfile)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
