% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createColorScheme.R
\name{qualitativeColours}
\alias{qualitativeColours}
\title{Create qualitative colours}
\source{
Modified based on https://gist.github.com/peterk87/6011397
}
\usage{
qualitativeColours(n, light = FALSE)
}
\arguments{
\item{n}{number of colors}

\item{light}{light colors TRUE or FALSE}
}
\value{
list of n different colors
}
\description{
Create qualitative colours
}
\examples{
\dontrun{
PhyloProfile:::qualitativeColours(5)
}
}
