% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxonomy_matrix.R
\name{rankIndexing}
\alias{rankIndexing}
\title{INDEXING ALL AVAILABLE RANKS (INCLUDING NORANK)}
\usage{
rankIndexing(rankListFile)
}
\arguments{
\item{rankListFile}{each row is a rank list of a taxon}
}
\value{
list of indexed ranks (dataframe)
}
\description{
INDEXING ALL AVAILABLE RANKS (INCLUDING NORANK)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
