% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rankList}
\alias{rankList}
\title{NCBI rank list for experimental data sets}
\format{
A data frame with up to 37 columns and 94 rows corresponding to 95
taxa in the 2 experimental data sets
}
\usage{
data(rankList)
}
\description{
Data frame, in which each row contains the complete taxonomy ranks from the
lowest systematic level (strain/species) upto the taxonomy root for one taxon
in the experimental data sets.
}
\keyword{datasets}
