% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{reduceProfile}
\alias{reduceProfile}
\title{Reduce the full processed profile into supertaxon level}
\usage{
reduceProfile(fullProfile)
}
\arguments{
\item{fullProfile}{dataframe contains the full processed profiles}
}
\value{
A reduced dataframe contains only profiles for the selected rank.
This dataframe contains only supertaxa and their value (%present, mVar1 &
mVar2) for each gene.
}
\description{
Reduce data of the phylogenetic profiles from input taxonomy
rank into supertaxon level (e.g. from species to phylum)
}
\examples{
data("fullProcessedProfile", package="phyloprofile")
reduceProfile(fullProcessedProfile)
}
\seealso{
\code{\link{parseInfoProfile}} for creating a full processed
profile dataframe, \code{\link{fullProcessedProfile}} for a demo full
processed profile dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
