% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDomainPlot.R
\name{resolveOverlapFeatures}
\alias{resolveOverlapFeatures}
\title{Modify domain dataframe to resolve overlapped domains/features}
\usage{
resolveOverlapFeatures(domainDf)
}
\arguments{
\item{domainDf}{input domain dataframe}
}
\value{
Domain dataframe with modified feature names that join multiple
domains of the same type that are not overlapped
}
\description{
Modify domain dataframe to resolve overlapped domains/features
}
\examples{
\dontrun{
NEED TO ADD AN EXAMPLE HERE
}
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
