% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_taxon_id.R
\name{search_taxonID_online}
\alias{search_taxonID_online}
\title{Search NCBI taxonomy IDs for a list of taxon names}
\usage{
search_taxonID_online(taxa)
}
\arguments{
\item{taxa}{list of taxon names}
}
\value{
dataframe contains NCBI taxonomy IDs for all input taxa
}
\description{
Search NCBI taxonomy IDs for a list of taxon names
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
