% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTaxonID.R
\name{searchTaxonIDOnline}
\alias{searchTaxonIDOnline}
\title{Search NCBI taxonomy IDs for a list of taxon names}
\usage{
searchTaxonIDOnline(taxa)
}
\arguments{
\item{taxa}{list of taxon names}
}
\value{
dataframe contains NCBI taxonomy IDs for all input taxa
}
\description{
Search NCBI taxonomy IDs for a list of taxon names
}
\examples{
taxa <- c("homo sapiens")
searchTaxonIDOnline(taxa)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
