% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{sortInputTaxa}
\alias{sortInputTaxa}
\title{Sort list of (super)taxa based on a selected reference (super)taxon}
\usage{
sortInputTaxa(taxonIDs, taxonNames, rankName, refTaxon, taxaTree)
}
\arguments{
\item{taxonIDs}{list of taxon IDs (e.g.: ncbi1234, ncbi9999, ...)}

\item{taxonNames}{list of taxon names and their corresponding taxonomy
ranks, IDs,...}

\item{rankName}{working taxonomy rank (e.g. "species", "phylum",...)}

\item{refTaxon}{selected reference taxon}

\item{taxaTree}{input taxonomy tree (optional)}
}
\value{
Taxonomy matrix for the input taxa ordered by the selected reference
taxon. This matrix is sorted either based on the NCBI taxonomy info, or
based on an user-defined taxonomy tree (if provided).
}
\description{
Sort list of (super)taxa based on a selected reference (super)taxon
}
\examples{
taxonIDs <- c(
    "ncbi272557", "ncbi176299", "ncbi3702", "ncbi876142", "ncbi9606"
)
taxonNames <- getInputTaxaName("species", taxonIDs)
sortInputTaxa(taxonIDs, taxonNames, "species", "Homo sapiens", NULL)
}
\seealso{
\code{\link{getNameList}}, \code{\link{getTaxonomyMatrix}},
\code{\link{createRootedTree}}, \code{\link{sortTaxaFromTree}},
\code{\link{getInputTaxaName}}, \code{\link{getInputTaxaID}},
\code{\link{createLongMatrix}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
