% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastaSeqs.R
\name{strReverse}
\alias{strReverse}
\title{Reverse string}
\usage{
strReverse(x)
}
\arguments{
\item{x}{input string}
}
\value{
a reversed string
}
\description{
Reverse string
}
\examples{
\dontrun{
strReverse("abc")
}
}
