% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxonomy_matrix.R
\name{taxa2dist}
\alias{taxa2dist}
\title{TAXA2DIST}
\usage{
taxa2dist(x, varstep = FALSE, check = TRUE, labels)
}
\arguments{
\item{x}{taxa matrix}

\item{varstep}{var-step}

\item{check}{check}

\item{labels}{labels}
}
\description{
TAXA2DIST
}
\author{
function from taxize library
}
