% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{taxaSelectGC}
\alias{taxaSelectGC}
\title{print list of available taxa}
\usage{
taxaSelectGC(rankSelectGC, inputTaxonID)
}
\arguments{
\item{rankSelectGC}{rank selected for group compariosn}

\item{inputTaxonID}{contains "seedID",  "orthoID", "feature", "start", "end"}
}
\value{
avilable taxa containing "ncbiID", "fullName", "rank", "parentID"
}
\description{
print list of available taxa
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
