% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonNamesReduced}
\alias{taxonNamesReduced}
\title{NCBI Taxonomy reduced data set}
\format{
A data frame with 4 columns:
\itemize{
    \item{ncbiID}{ e.g. "10090"}
    \item{fullName}{ e.g. "Mus musculus"}
    \item{rank}{ e.g. "species"}
    \item{parentID}{ e.g. "862507"}
}
}
\usage{
data(taxonNamesReduced)
}
\description{
A list of NCBI taxonomy info (including taxon IDs, taxon names, their
systematic taxonomy rank and IDs of their next rank - parent IDs) for 95 taxa
in two experimental sets included in PhyloProfilData package.
}
\keyword{datasets}
