% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxonomy_matrix.R
\name{taxonomy_table_creator}
\alias{taxonomy_table_creator}
\title{Align ncbi taxonomy IDs of all taxa into a sorted rank list}
\usage{
taxonomy_table_creator(idListFile, rankListFile)
}
\arguments{
\item{idListFile}{a text file whose each row is a rank+ID list of a taxon}

\item{rankListFile}{a text file whose each row is a rank list of a taxon}
}
\value{
aligned taxonomy matrix  (taxonomyMatrix.txt)
}
\description{
Align ncbi taxonomy IDs of all taxa into a sorted rank list
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
