% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomyMatrix}
\alias{taxonomyMatrix}
\title{Taxonomy matrix for experimental data sets}
\format{
A data frame with up to 147 columns and 95 rows corresponding to 95
taxa in the 2 experimental data sets
}
\usage{
data(taxonomyMatrix)
}
\description{
Data frame containing the fully aligned taxonomy IDs of 95 taxa in the
experimental data sets. By talking into account both the defined ranks (e.g.
strain, This data is used for clustering and then creating a
taxon tree. It is used also for cross-linking between different taxonomy
ranks within a taxon.
}
\keyword{datasets}
