% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapClustering.R
\name{umapClustering}
\alias{umapClustering}
\title{Perform UMAP clustering}
\usage{
umapClustering(data4umap = NULL, by = "taxa", type = "binary", 
    randomSeed = 123)
}
\arguments{
\item{data4umap}{data for UMAP clustering (output from prepareUmapData)}

\item{by}{cluster data by "taxa" (default) or "genes"}

\item{type}{type of data, either "binary" (default) or "non-binary"}

\item{randomSeed}{random seed. Default: 123}
}
\value{
A list contain UMAP cluster objects
}
\description{
Perform UMAP clustering
}
\examples{
rawInput <- system.file(
   "extdata", "test.main.long", package = "PhyloProfile", mustWork = TRUE
)
longDf <- createLongMatrix(rawInput)
data4umap <- prepareUmapData(longDf, "phylum")
umapClustering(data4umap)
}
\seealso{
\code{\link{prepareUmapData}}
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
