% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{calcPresSpec}
\alias{calcPresSpec}
\title{Calculate percentage of present species in each ortholog group}
\usage{
calcPresSpec(profileWithTax, taxaCount)
}
\arguments{
\item{profileWithTax}{long data frame of main phyloprofile input together
with their taxonomy info}

\item{taxaCount}{number of species occur in each supertaxon}
}
\value{
A data frame with % of present species for each seed protein in
each supertaxon
}
\description{
Calculate percentage of present species in each ortholog group
}
\examples{
# NOTE: for internal testing only - not recommended for outside using
data("profileWithTaxonomy", package="phyloprofile")
taxaCount <- plyr::count(profileWithTaxonomy, "supertaxon")
taxaCount$freq <- 1
calcPresSpec(profileWithTaxonomy, taxaCount)
}
\seealso{
\code{\link{profileWithTaxonomy}} for a demo input data
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
