% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{checkOmaID}
\alias{checkOmaID}
\title{Check the validity of input OMA IDs}
\usage{
checkOmaID(ids)
}
\arguments{
\item{ids}{list of ids needs to be checked}
}
\value{
List of invalid IDs (not readable for OMA)
}
\description{
Check if input IDs are valid OMA IDs for OMA Browser
}
\examples{
print("Uncomment the following line to run the function")
# checkOmaID("HUMAN29398")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
