% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDomainPlot.R
\name{checkOverlapDomains}
\alias{checkOverlapDomains}
\title{Identify feature type(s) containing overlapped domains/features}
\usage{
checkOverlapDomains(domainDf)
}
\arguments{
\item{domainDf}{input domain dataframe}
}
\value{
List of feature types that have overlapped domains
}
\description{
Identify feature type(s) containing overlapped domains/features
}
\examples{
\dontrun{
NEED TO ADD AN EXAMPLE HERE
}
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
