% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterProfile.R
\name{clusterDataDend}
\alias{clusterDataDend}
\title{Create a dendrogram tree from the distance matrix}
\usage{
clusterDataDend(distanceMatrix, clusterMethod)
}
\arguments{
\item{distanceMatrix}{calculated distance matrix}

\item{clusterMethod}{clustering method ("single", "complete",
"average" for UPGMA, "mcquitty" for WPGMA, "median" for WPGMC,
or "centroid" for UPGMC)}
}
\value{
A dendrogram tree object
}
\description{
Create a dendrogram tree from the distance matrix
}
\examples{
data("fullProcessedProfileLarge", package="PhyloProfile")
data <- fullProcessedProfileLarge
profileType <- "binary"
profiles <- getDataClustering(
    data, profileType, var1AggregateBy, var2AggregateBy)
distMethod <- "mutualInformation"
distanceMatrix <- getDistanceMatrix(profiles, distMethod)
clusterMethod <- "complete"
clusterDataDend(distanceMatrix, clusterMethod)
}
\seealso{
\code{\link{getDataClustering}},
\code{\link{getDistanceMatrix}}, \code{\link{hclust}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
