% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDomainPlot.R
\name{createArchiPlot}
\alias{createArchiPlot}
\title{Create protein's domain architecure plot}
\usage{
createArchiPlot(info, domainDf, labelArchiSize, titleArchiSize)
}
\arguments{
\item{info}{a list contains seed and orthologs IDs}

\item{domainDf}{dataframe contains domain info}

\item{labelArchiSize}{lable size (in px)}

\item{titleArchiSize}{title size (in px)}
}
\value{
A plot as arrangeGrob object. Use grid::grid.draw(plot) to render.
}
\description{
Create architecture plot for both seed and orthologous protein.
If domains of ortholog are missing, only architecture of seed protein will
be plotted. NOTE: seed protein ID is the one being shown in the profile plot,
which normally is also the orthologous group ID.
}
\examples{
seedID <- "OG_1009"
orthoID <- "A.thaliana@3702@241736"
info <- c(seedID, orthoID)
domainFile <- system.file(
    "extdata", "domainFiles/OG_1009.domains",
    package = "phyloprofile", mustWork = TRUE
)
domainDf <- parseDomainInput(seedID, domainFile, "file")
plot <- createArchiPlot(info, domainDf, 9, 9)
grid::grid.draw(plot)
}
\seealso{
\code{\link{domainPlotting}}, \code{\link{sortDomains}},
\code{\link{parseDomainInput}}, \code{\link{getQualColForVector}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
