% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGeneAge.R
\name{createGeneAgePlot}
\alias{createGeneAgePlot}
\title{Create gene age plot}
\usage{
createGeneAgePlot(countDf, geneAgeText)
}
\arguments{
\item{countDf}{data for plotting gene age}

\item{geneAgeText}{text size}
}
\value{
A gene age distribution plot as a ggplot2 object
}
\description{
Create gene age plot
}
\examples{
countDf <- data.frame(
    age = "OG_1017",
    frea = 2,
    percentage = 100,
    pos = 50
)
geneAgeText <- 1
createGeneAgePlot(countDf, geneAgeText)
}
\seealso{
\code{\link{estimateGeneAge}} and \code{\link{geneAgePlotDf}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
