% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDistribution.R
\name{createPercentageDistributionData}
\alias{createPercentageDistributionData}
\title{Create data for percentage present taxa distribution}
\usage{
createPercentageDistributionData(inputData, rankName)
}
\arguments{
\item{inputData}{dataframe contains raw input data in long format}

\item{rankName}{name of the working taxonomy rank (e.g. "species", "family")}
}
\value{
A dataframe ready for analysing the distribution of the percentage of
species in the selected supertaxa
}
\description{
Create data for percentage present taxa distribution
}
\examples{
data("mainLongRaw", package="PhyloProfile")
createPercentageDistributionData(mainLongRaw, "class")
}
\seealso{
\code{\link{mainLongRaw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
