% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{createProfileFromOma}
\alias{createProfileFromOma}
\title{Create a phylogenetic profile from a raw OMA dataframe}
\usage{
createProfileFromOma(finalOmaDf = NULL)
}
\arguments{
\item{finalOmaDf}{raw OMA data for a list of proteins (see ?getDataForOneOma)}
}
\value{
Dataframe of the phylogenetic profiles in long format, which
contains the seed protein IDs, their orthologous proteins and the
corresponding taxononmy IDs of the orthologs.
}
\description{
Create a phylogenetic profile from a raw OMA dataframe
}
\examples{
omaData <- getDataForOneOma("HUMAN29397", "OG")
createProfileFromOma(omaData)
}
\seealso{
\code{\link{getDataForOneOma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
