% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDistribution.R
\name{createVarDistPlot}
\alias{createVarDistPlot}
\title{Create distribution plot}
\usage{
createVarDistPlot(data, varName, varType, percent, distTextSize)
}
\arguments{
\item{data}{dataframe contains data for plotting}

\item{varName}{name of the variable that need to be analyzed (either name of
variable 1 or variable 2 or "percentage of present taxa")}

\item{varType}{type of variable (either "var1", "var2" or "presSpec")}

\item{percent}{range of percentage cutoff}

\item{distTextSize}{text size of the distribution plot}
}
\value{
A distribution plot as a ggplot object
}
\description{
Create distribution plot for one of the additional variable or
the percentage of the species present in the supertaxa.
}
\examples{
data("mainLongRaw", package="PhyloProfile")
data <- createVariableDistributionData(
    mainLongRaw, 0, 1, 0.5, 1
)
varName <- "Variable abc"
varType <- "var1"
percent <- c(0,1)
distTextSize <- 12
createVarDistPlot(
    data,
    varName,
    varType,
    percent,
    distTextSize
)
}
\seealso{
\code{\link{mainLongRaw}},
\code{\link{createVariableDistributionData}},
\code{\link{createVariableDistributionDataSubset}},
\code{\link{createPercentageDistributionData}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
