% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{dataFeatureTaxGroup}
\alias{dataFeatureTaxGroup}
\title{Create data for feature distribution comparison plot}
\usage{
dataFeatureTaxGroup(mainDf, domainDf, inGroup, gene, featureThreshold)
}
\arguments{
\item{mainDf}{input phylogenetic profile in long format (see ?mainLongRaw 
and ?createLongMatrix)}

\item{domainDf}{dataframe contains domain info for the seed and ortholog. 
This including the seed ID, orthologs IDs, sequence lengths, feature names, 
start and end positions, feature weights (optional) and the status to 
determine if that feature is important for comparison the architecture 
between 2 proteins* (e.g. seed protein vs ortholog) (optional). (see 
?parseDomainInput)}

\item{inGroup}{ID list of in-group taxa (e.g. "ncbi1234")}

\item{gene}{ID of gene that need to be plotted the feature distribution 
comparison between in- and out-group taxa.}

\item{featureThreshold}{cutoff for occurence frequencies of the protein 
feature instances in each taxon group (in percentage). Default = 0.}
}
\value{
Dataframe containing all feature names, their frequencies (absolute 
count and percentage) in each taxon group and the corresponding taxa group.
}
\description{
Create data for plotting the distribution of the protein domain 
features between 2 group of taxa for a selected gene.
}
\examples{
data("mainLongRaw", package="PhyloProfile")
data <- mainLongRaw
gene <- "OG_1017"
inputFile <- system.file(
    "extdata", "domainFiles/OG_1017.domains",
    package = "PhyloProfile", mustWork = TRUE
)
type <- "file"
domainDf <- parseDomainInput(gene, inputFile, type)
inGroup <- c("ncbi876142", "ncbi586133")
featureThreshold <- 15
dataFeatureTaxGroup(data, domainDf, inGroup, gene, featureThreshold)
}
\seealso{
\code{\link{createLongMatrix}}, \code{\link{parseDomainInput}}
}
\author{
Vinh Tran (tran@bio.uni-frankfurt.de)
}
