% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGeneAge.R
\name{geneAgePlotDf}
\alias{geneAgePlotDf}
\title{Create data for plotting gene ages}
\usage{
geneAgePlotDf(geneAgeDf)
}
\arguments{
\item{geneAgeDf}{data frame containing estimated gene ages for seed proteins}
}
\value{
A dataframe for plotting gene age plot containing the absolute number
and percentage of genes for each calculated evolutionary ages and the 
corresponding position for writting those number on the plot.
}
\description{
Create data for plotting gene ages
}
\examples{
geneAgeDf <- data.frame(
geneID = c("OG_1017", "OG_1019"),
cat = c("0000001", "0000001"),
age = c("07_LUCA", "07_LUCA")
)
geneAgePlotDf(geneAgeDf)
}
\seealso{
\code{\link{estimateGeneAge}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
