% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{getAllDomainsOma}
\alias{getAllDomainsOma}
\title{Create domain annotation dataframe from a raw OMA dataframe}
\usage{
getAllDomainsOma(finalOmaDf = NULL)
}
\arguments{
\item{finalOmaDf}{raw OMA data for a list of proteins (see ?getDataForOneOma)}
}
\value{
Dataframe of the domain annotation used for PhyloProfile, which
contains seed IDs, ortholog IDs, ortholog lengths, annotated features, start
and end positions of those features.
}
\description{
Create domain annotation dataframe from a raw OMA dataframe
}
\examples{
print("Uncomment the following line to run the function")
# omaData <- getDataForOneOma("HUMAN29397", "OG")
# getAllDomainsOma(omaData)
}
\seealso{
\code{\link{getDataForOneOma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
