% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{getAllFastaOma}
\alias{getAllFastaOma}
\title{Get all fasta sequences from a raw OMA dataframe}
\usage{
getAllFastaOma(finalOmaDf = NULL)
}
\arguments{
\item{finalOmaDf}{raw OMA data for a list of proteins (see ?getDataForOneOma)}
}
\value{
A list contains all protein sequences in fasta format.
}
\description{
Get all fasta sequences from a raw OMA dataframe
}
\examples{
omaData <- getDataForOneOma("HUMAN29397", "OG")
getAllFastaOma(omaData)
}
\seealso{
\code{\link{getDataForOneOma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
