% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getCommonAncestor}
\alias{getCommonAncestor}
\title{Get all taxa that share a common ancestor}
\usage{
getCommonAncestor(inputTaxa = NULL, inGroup = NULL)
}
\arguments{
\item{inputTaxa}{ID list of all input taxa (e.g. "ncbi12345")}

\item{inGroup}{ID list of selected taxa used for identify the common ancestor
(e.g.: "ncbi55555")}
}
\value{
A list containing the taxonomy rank and name of the common ancestor,
together with a dataframe storing the full taxonomy info of all taxa that
share that corresponding common ancestor.
}
\description{
Identify the common ancestor for a selected taxa and return a
list of all taxa that have that common ancestor from an large input taxa set.
}
\examples{
inputTaxa <- c("ncbi103449", "ncbi1288291", "ncbi278021", "ncbi27973",
    "ncbi31281", "ncbi40302", "ncbi586133", "ncbi6035", "ncbi70536",
    "ncbi876142", "ncbi993615")
inGroup <-  c("ncbi27973", "ncbi6035")
getCommonAncestor(inputTaxa, inGroup)
}
\author{
Vinh Tran (tran@bio.uni-frankfurt.de)
}
