% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getCommonAncestor}
\alias{getCommonAncestor}
\title{Generate the inGroup}
\usage{
getCommonAncestor(inGroup, rank, nameList, selectedInGroup,
    ncbiIDList)
}
\arguments{
\item{inGroup}{list of taxa}

\item{rank}{selected rank}

\item{nameList}{contains "ncbiID", "fullName", "rank", "parentID"}

\item{selectedInGroup}{contains "abbrName, "ncbiID", fullName", "strain",
"genus"..}

\item{ncbiIDList}{list of input taxon IDs}
}
\value{
common anchestor
}
\description{
Generate the inGroup
}
\examples{
inGroup <- c("Aeropyrum pernix", "Agrobacterium fabrum")
rank <- "species"
nameListFile <- system.file(
    "extdata", "data/taxonNamesFull.txt",
    package = "phyloprofile", mustWork = TRUE
)
nameList <- as.data.frame(data.table::fread(nameListFile))
selectedInGroup <- getTaxonomyMatrix(FALSE, NULL)
ncbiIDList <- c(56636, 1176649)
getCommonAncestor(
    inGroup,
    rank,
    nameList,
    selectedInGroup,
    ncbiIDList
)
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
