% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyCoreGene.R
\name{getCoreGene}
\alias{getCoreGene}
\title{Identify core genes for a list of selected taxa}
\usage{
getCoreGene(rankName, taxaCore, processedProfileData,
    var1Cutoff, var2Cutoff, percentCutoff, coreCoverage)
}
\arguments{
\item{rankName}{working taxonomy rank (e.g. "species", "genus", "family")}

\item{taxaCore}{list of selected taxon names}

\item{processedProfileData}{dataframe contains the full processed
phylogenetic profiles (see ?fullProcessedProfile or ?parseInfoProfile)}

\item{var1Cutoff}{cutoff for var1}

\item{var2Cutoff}{cutoff for var2}

\item{percentCutoff}{cutoff for percentage of species present in each
supertaxon}

\item{coreCoverage}{the least number of selected taxa should be considered}
}
\value{
A list of identified core genes.
}
\description{
Identify core genes for a list of selected (super)taxa. The 
identified core genes must be present in at least a certain proportion of 
species in each selected (super)taxon (identified via percentCutoff) and 
that criteria must be fullfilled for a certain percentage of selected taxa 
or all of them (determined via coreCoverage).
}
\examples{
data("fullProcessedProfile", package="PhyloProfile")
rankName <- "class"
taxaCore <- c("Mammalia", "Mucorales", "Alphaproteobacteria")
processedProfileData <- fullProcessedProfile
var1Cutoff <- c(0.75, 1.0)
var2Cutoff <- c(0.75, 1.0)
percentCutoff <- c(0.0, 1.0)
coreCoverage <- 1
getCoreGene(
    rankName,
    taxaCore,
    processedProfileData,
    var1Cutoff, var2Cutoff,
    percentCutoff, coreCoverage
)
}
\seealso{
\code{\link{parseInfoProfile}} for creating a full processed
profile dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
