% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastaSeqs.R
\name{getFastaSeqs}
\alias{getFastaSeqs}
\title{Get fasta sequences}
\usage{
getFastaSeqs(dataIn, filein, demoData, inputTypeFasta,
    concatFasta, path, dirFormat, fileExt, idFormat)
}
\arguments{
\item{dataIn}{a dataframe of the input phylogenetic profiles, that contains
at least 3 columns: geneIDs, orthoIDs and ncbiIDs (see ?mainLongRaw or 
?fromInputToProfile)}

\item{filein}{raw phylogenetic profile input file (for checking input type)}

\item{demoData}{name of demo data set (either "ampk-tor", "lca-micros",
or "none")}

\item{inputTypeFasta}{source type of fasta sequences ("Concatenated fasta 
file" or "Fasta folder")}

\item{concatFasta}{input concatenated fasta file (NULL if not available)}

\item{path}{path to fasta folder (NULL if concatFasta already given)}

\item{dirFormat}{directory format (either 1 for "path/speciesID.fa*" or 2 for
"path/speciesID/speciesID.fa*") (NULL if concatFasta already given)}

\item{fileExt}{fasta file extension ("fa", "fasta", "fas" or "txt") (NULL if 
concatFasta already given)}

\item{idFormat}{fasta header format (1 for ">speciesID:seqID", 2 for
">speciesID@seqID", 3 for ">speciesID|seqID" or 4 for "seqID") (NULL if 
concatFasta already given)}
}
\value{
A dataframe with one column contains sequences in fasta format.
}
\description{
Get fasta sequences for the input phylogenetic profiles.
}
\examples{
data("mainLongRaw", package="PhyloProfile")
dataIn <- mainLongRaw
filein <- system.file(
    "extdata", "test.main.long", package = "PhyloProfile", mustWork = TRUE
)
demoData <- "none"
inputTypeFasta <- "Concatenated fasta file"
concatFasta <- system.file(
    "extdata", "fastaFiles/concatenatedFile.fa",
    package = "PhyloProfile", mustWork = TRUE
)
# the following parameters are conly required if upload a fasta folder!
path <- NULL
dirFormat <- NULL
fileExt <- NULL
idFormat <- NULL
# get fasta sequences
getFastaSeqs(
    dataIn, filein, demoData,
    inputTypeFasta,
    concatFasta,
    path,
    dirFormat,
    fileExt,
    idFormat
)
}
\seealso{
\code{\link{checkInputValidity}}, \code{\link{mainLongRaw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
