% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmaBrowser.R
\name{getOmaDataForOneOrtholog}
\alias{getOmaDataForOneOrtholog}
\title{Get taxonomy ID, sequence and annotation for one OMA sequence}
\usage{
getOmaDataForOneOrtholog(id)
}
\arguments{
\item{id}{oma ID of a ortholog}
}
\value{
data frame
}
\description{
Get taxonomy ID, sequence and annotation for one OMA sequence
}
\examples{
getOmaDataForOneOrtholog("HUMAN29397")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
