% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{getSelectedTaxonNames}
\alias{getSelectedTaxonNames}
\title{Get a subset of input taxa based on a selected taxonomy rank}
\usage{
getSelectedTaxonNames(inputTaxonIDs, rank, higherRank, higherID, 
    higherName)
}
\arguments{
\item{inputTaxonIDs}{list of input taxon IDs (e.g. 876142)}

\item{rank}{taxonomy rank of input taxa (e.g. "species")}

\item{higherRank}{selected taxonomy rank (e.g. "phylum")}

\item{higherID}{supertaxon ID (e.g. 6029). NOTE: either supertaxon ID or 
name is required, not neccessary to give both.}

\item{higherName}{supertaxon name (e.g. "Microsporidia"). NOTE: either 
supertaxon ID or name is required, not neccessary to give both.}
}
\value{
A list of taxa from the input taxon list that belong to the selected
supertaxon.
}
\description{
Get a subset of taxon names from an input list of taxa based on
a selected supertaxon (identified by its taxonomy rank and supertaxon name).
}
\examples{
getNameList()
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
