% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTaxaGroups.R
\name{getSignificantGenes}
\alias{getSignificantGenes}
\title{Get the dataframe with the significant genes}
\usage{
getSignificantGenes(inGroup, selectedGenesList, rank, var,
    useCommonAncestor, referenceTaxon, parameters, ncbiIDList,
    dataFull, rightFormatFeatures, domains)
}
\arguments{
\item{inGroup}{list of taxa}

\item{selectedGenesList}{list of genes}

\item{rank}{selected taxonamy rank}

\item{var}{variable for which to  calculate the significance}

\item{useCommonAncestor}{boolean if the common anchestor should be used}

\item{referenceTaxon}{taxon which is used as reference}

\item{parameters}{contains "showPValue","highlightSignificant",
"significance", "var1ID", "var2ID", "xSizeGC", "ySizeGC",
"interestingFeatures", "angleGC", "legendGC", "legendSizeGC"}

\item{ncbiIDList}{list of ncbi ids}

\item{dataFull}{full processed main data}

\item{rightFormatFeatures}{boolean if the features have the right format}

\item{domains}{dataframe holding the domain input}
}
\value{
dataframe with the significant genes
}
\description{
Get the dataframe with the significant genes
}
\examples{
data("mainLongRaw", package="phyloprofile")
data("fullProcessedProfile", package="phyloprofile")
inGroup <- c("Aeropyrum pernix", "Agrobacterium fabrum")
selectedGenesList <- "OG_1017"
rank <- "species"
var <- ""
useCommonAncestor <- FALSE
referenceTaxon <- "Homo sapiens"
ncbiIDList <- getInputTaxaID(mainLongRaw)
dataFull <- fullProcessedProfile
rightFormatFeatures <- TRUE
domainFile <- system.file(
    "extdata", "domainFiles/OG_1017.domains",
    package = "phyloprofile", mustWork = TRUE
)
domains <- parseDomainInput("OG_1017", domainFile, "file")
parameters <- list(
    "showPValue" = TRUE,
    "highlightSignificant" = FALSE,
    "significance" = 0.05,
    "var1ID" = "1st variable",
    "var2ID" = "2nd variable",
    "xSizeGC" = 9,
    "ySizeGC" = 9,
    "interestingFeatures" = NULL,
    "angleGC" = 90,
    "legendGC" = "none",
    "legendSizeGC" = 9,
    "pValuesSize" = 9
)
getSignificantGenes(
    inGroup,
    selectedGenesList,
    rank,
    var,
    useCommonAncestor,
    referenceTaxon,
    parameters,
    ncbiIDList,
    dataFull,
    rightFormatFeatures,
    domains
)
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
