% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProfilePlot.R
\name{highlightProfilePlot}
\alias{highlightProfilePlot}
\title{Highlight gene and/or taxon of interest on the phylogenetic profile plot}
\usage{
highlightProfilePlot(data, plotParameter = NULL, taxonHighlight =
    "none", rankName = "none", geneHighlight = "none")
}
\arguments{
\item{data}{dataframe for plotting the heatmap phylogentic profile (either
full or subset profiles)}

\item{plotParameter}{plot parameters, including (1) type of x-axis "taxa" or
"genes" - default = "taxa"; (2+3) names of 2 variables var1ID and var2ID -
default = "var1" & "var2"; (4) color for lowest var1 - default = "#FF8C00";
(5) color for highest var1 - default = "#4682B4"; (6) color for lowest var2 -
default = "#FFFFFF", (7) color for highest var2 - default = "#F0E68C", (8)
color of co-orthologs - default = "#07D000"; (9+10+11) text sizes for x, y
axis and legend - default = 9 for each; (12) legend position "top", "bottom",
"right", "left" or "none" - default = "top"; (13) zoom ratio of the
co-ortholog dots from -1 to 3 - default = 0; (14) angle of x-axis from 0 to
90 - default = 60; (14) show/hide separate line for reference taxon 1/0 -
default = 0; (15) enable/disable coloring gene categories TRUE/FALSE -
default = FALSE). NOTE: Leave blank or NULL to use default values.}

\item{taxonHighlight}{taxon of interst. Default = "none".}

\item{rankName}{working taxonomy rank (needed only for highlight taxon).}

\item{geneHighlight}{gene of interest. Default = "none".}
}
\value{
A profile heatmap plot with highlighted gene and/or taxon of interest
as ggplot object.
}
\description{
Highlight gene and/or taxon of interest on the phylogenetic profile plot
}
\examples{
data("superTaxonProfile", package="PhyloProfile")
plotDf <- dataMainPlot(superTaxonProfile)
plotParameter <- list(
    "xAxis" = "taxa",
    "var1ID" = "FAS_FW",
    "var2ID"  = "FAS_BW",
    "lowColorVar1" =  "#FF8C00",
    "highColorVar1" = "#4682B4",
    "lowColorVar2" = "#FFFFFF",
    "highColorVar2" = "#F0E68C",
    "paraColor" = "#07D000",
    "xSize" = 8,
    "ySize" = 8,
    "legendSize" = 8,
    "mainLegend" = "top",
    "dotZoom" = 0,
    "xAngle" = 60,
    "guideline" = 0,
    "colorByGroup" = FALSE
)
taxonHighlight <- "none"
rankName <- "class"
geneHighlight <- "100265at6656"
highlightProfilePlot(
    plotDf, plotParameter, taxonHighlight, rankName, geneHighlight
)
}
\seealso{
\code{\link{dataMainPlot}}, \code{\link{dataCustomizedPlot}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
