% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProfilePlot.R
\name{highlightProfilePlot}
\alias{highlightProfilePlot}
\title{Highlight gene and/or taxon of interest on the phylogenetic profile plot}
\usage{
highlightProfilePlot(data, plotParameter = NULL, taxonHighlight =
    "none", rankName = "none", geneHighlight = "none")
}
\arguments{
\item{data}{dataframe for plotting the heatmap phylogentic profile (either
full or subset profiles)}

\item{plotParameter}{plot parameters, including (1) type of x-axis "taxa" or
"genes" - default = "taxa"; (2+3) names of 2 variables var1ID and var2ID -
default = "var1" & "var2"; (4+5) mid value and color for mid value of var1 -
default is 0.5 and #FFFFFF; (6) color for lowest var1 - default = "#FF8C00";
(7) color for highest var1 - default = "#4682B4"; (8+9) mid value and color
for mid value of var2 - default is 1 and #FFFFFF;(10) color for lowest var2 -
default = "#FFFFFF", (11) color for highest var2 - default = "#F0E68C", (12)
color of co-orthologs - default = "#07D000"; (13+14+15) text sizes for x, y
axis and legend - default = 9 for each; (16) legend position "top", "bottom",
"right", "left" or "none" - default = "top"; (17) zoom ratio of the
co-ortholog dots from -1 to 3 - default = 0; (18) angle of x-axis from 0 to
90 - default = 60; (19) show/hide separate line for reference taxon 1/0 -
default = 0; (20) enable/disable coloring gene categories TRUE/FALSE -
default = FALSE). NOTE: Leave blank or NULL to use default values.}

\item{taxonHighlight}{taxon of interst. Default = "none".}

\item{rankName}{working taxonomy rank (needed only for highlight taxon).}

\item{geneHighlight}{gene of interest. Default = "none".}
}
\value{
A profile heatmap plot with highlighted gene and/or taxon of interest
as ggplot object.
}
\description{
Highlight gene and/or taxon of interest on the phylogenetic profile plot
}
\examples{
data("finalProcessedProfile", package="PhyloProfile")
plotDf <- dataMainPlot(finalProcessedProfile)
plotParameter <- list(
    "xAxis" = "taxa",
    "var1ID" = "FAS_FW",
    "var2ID"  = "FAS_BW",
    "midVar1" = 0.5,
    "midColorVar1" =  "#FFFFFF",
    "lowColorVar1" =  "#FF8C00",
    "highColorVar1" = "#4682B4",
    "midVar2" = 1,
    "midColorVar2" =  "#FFFFFF",
    "lowColorVar2" = "#CB4C4E",
    "highColorVar2" = "#3E436F",
    "paraColor" = "#07D000",
    "xSize" = 8,
    "ySize" = 8,
    "legendSize" = 8,
    "mainLegend" = "top",
    "dotZoom" = 0,
    "xAngle" = 60,
    "guideline" = 0,
    "colorByGroup" = FALSE,
    "colorByOrthoID" = FALSE
)
taxonHighlight <- "none"
rankName <- "class"
geneHighlight <- "100265at6656"
highlightProfilePlot(
    plotDf, plotParameter, taxonHighlight, rankName, geneHighlight
)
}
\seealso{
\code{\link{dataMainPlot}}, \code{\link{dataCustomizedPlot}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
