% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDomainPlot.R
\name{joinPlotMergeLegends}
\alias{joinPlotMergeLegends}
\title{Join multiple plots and merge legends}
\usage{
joinPlotMergeLegends(
  df1 = NULL,
  df2 = NULL,
  plot1 = NULL,
  plot2 = NULL,
  position = c("bottom", "right"),
  font = "Arial"
)
}
\arguments{
\item{df1}{Data frame for plot 1}

\item{df2}{Data frame for plot 2}

\item{plot1}{ggplot object of plot 1}

\item{plot2}{ggplot object of plot 2}

\item{position}{position of legend (bottom or right)}

\item{font}{font of text}
}
\value{
joined plots with merged legend as a grid object
}
\description{
Join multiple plots and merge legends
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
