% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDomainPlot.R
\name{pairDomainPlotting}
\alias{pairDomainPlotting}
\title{Create architecure plot for a pair of seed and ortholog protein}
\usage{
pairDomainPlotting(seed, ortho, seedDf, orthoDf, minStart, maxEnd, 
    labelSize, titleSize)
}
\arguments{
\item{seed}{Seed ID}

\item{ortho}{Ortho ID}

\item{seedDf}{domain dataframe for seed domains containing the seed ID, 
ortholog ID, sequence length, feature names, start and end positions, 
feature weights (optional) and the status to determine if that feature is 
important for comparison the architecture between 2 proteins* (e.g. seed 
protein vs ortholog) (optional).}

\item{orthoDf}{domain dataframe for ortholog domains (same format as seedDf).}

\item{minStart}{the smallest start position of all domains}

\item{maxEnd}{the highest stop position of all domains}

\item{labelSize}{lable size. Default = 12.}

\item{titleSize}{title size. Default = 12.}
}
\value{
Domain plot of a pair proteins as a arrangeGrob object.
}
\description{
Create architecure plot for a pair of seed and ortholog protein
}
\examples{
\dontrun{
seed <- "101621at6656"
ortho <- "101621at6656|AGRPL@224129@0|224129_0:001955|1"
ortho <- gsub("\\\\|", ":", ortho)
grepID <- paste(seed, "#", ortho, sep = "")
domainFile <- system.file(
    "extdata", "domainFiles/101621at6656.domains",
    package = "PhyloProfile", mustWork = TRUE
)
domainDf <- parseDomainInput(seed, domainFile, "file")
subdomainDf <- domainDf[grep(grepID, domainDf$seedID), ]
subdomainDf$feature <- as.character(subdomainDf$feature)
orthoDf <- subdomainDf[subdomainDf$orthoID == ortho,]
seedDf <- subdomainDf[subdomainDf$orthoID != ortho,]
minStart <- min(subdomainDf$start)
maxEnd <- max(c(subdomainDf$end, subdomainDf$length))
g <- pairDomainPlotting(seed,ortho,seedDf,orthoDf,minStart,maxEnd,9,9)    
grid::grid.draw(g)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
