% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppTree}
\alias{ppTree}
\title{An example of a taxonomy tree in newick format.}
\format{
A data frame with only one entry
\describe{
    \item{V1}{ tree in newick format}
}
}
\usage{
data(ppTree)
}
\description{
An example of a taxonomy tree in newick format.
}
