% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTaxonomyMatrix.R
\name{rankIndexing}
\alias{rankIndexing}
\title{Indexing all available ranks (including norank)}
\usage{
rankIndexing(rankListFile)
}
\arguments{
\item{rankListFile}{Input file, where each row is a rank list of a taxon}
}
\value{
A dataframe containing a list of indexed ranks
}
\description{
Indexing all available ranks (including norank)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
rankListFile <- system.file(
    "extdata", "data/rankList.txt", package = "phyloprofile", mustWork = TRUE
)
rankIndexing(rankListFile)
}
