% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhyloprofile.R
\name{runPhyloProfile}
\alias{runPhyloProfile}
\title{Run PhyloProfile app}
\usage{
runPhyloProfile()
}
\value{
shiny application of PhyloProfile
}
\description{
Run PhyloProfile app
}
\examples{
?PhyloProfile::runPhyloProfile
\dontrun{
runPhyloProfile()
}
}
