% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhyloprofile.R
\name{runPhyloprofile}
\alias{runPhyloprofile}
\title{Run PhyloProfile app}
\usage{
runPhyloprofile()
}
\value{
shiny application of PhyloProfile
}
\description{
Run PhyloProfile app
}
\examples{
?phyloprofile::runPhyloprofile
\dontrun{
runPhyloprofile()
}
}
