% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDomainPlot.R
\name{sortDomainsByList}
\alias{sortDomainsByList}
\title{Sort one domain dataframe based on list of ordered feature types}
\usage{
sortDomainsByList(domainDf = NULL, featureTypeOrder = NULL)
}
\arguments{
\item{domainDf}{domain dataframe}

\item{featureTypeOrder}{vector of ordered feature types}
}
\value{
Dataframe contains sorted domain list.
}
\description{
Sort domain dataframe of one protein based on a given list of
ordered feature types
}
\examples{
\dontrun{
NEED TO ADD AN EXAMPLE HERE
}
}
\author{
Vinh Tran tran@bio.uni-frankfurt.de
}
