% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhylotree.R
\name{taxa2dist}
\alias{taxa2dist}
\title{taxa2dist}
\usage{
taxa2dist(x, varstep = FALSE, check = TRUE, labels)
}
\arguments{
\item{x}{taxa matrix}

\item{varstep}{var-step}

\item{check}{check}

\item{labels}{labels}
}
\value{
a distance matrix
}
\description{
taxa2dist
}
\author{
function from taxize library
}
