% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxonomy_matrix.R
\name{taxonomy_table_creator}
\alias{taxonomy_table_creator}
\title{ARRANGE RANK IDs INTO SORTED RANK LIST (index2RankDf)}
\usage{
taxonomy_table_creator(idListFile, rankListFile)
}
\arguments{
\item{idListFile}{each row is a rank+ID list of a taxon}

\item{rankListFile}{each row is a rank list of a taxon}
}
\value{
aligned taxonomy matrix
}
\description{
ARRANGE RANK IDs INTO SORTED RANK LIST (index2RankDf)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
