% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTaxonomyMatrix.R
\name{taxonomyTableCreator}
\alias{taxonomyTableCreator}
\title{Align ncbi taxonomy IDs of all taxa into a sorted rank list}
\usage{
taxonomyTableCreator(idListFile, rankListFile)
}
\arguments{
\item{idListFile}{a text file whose each row is a rank+ID list of a taxon}

\item{rankListFile}{a text file whose each row is a rank list of a taxon}
}
\value{
An aligned taxonomy dataframe
}
\description{
Align ncbi taxonomy IDs of all taxa into a sorted rank list
}
\examples{
idListFile <- system.file(
    "extdata", "data/idList.txt", package = "PhyloProfile", mustWork = TRUE
)
rankListFile <- system.file(
    "extdata", "data/rankList.txt", package = "PhyloProfile", mustWork = TRUE
)
taxonomyTableCreator(idListFile, rankListFile)
}
\seealso{
\code{\link{rankIndexing}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
