% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMainInput.R
\name{xmlParser}
\alias{xmlParser}
\title{Parse orthoXML input file}
\usage{
xmlParser(inputFile)
}
\arguments{
\item{inputFile}{input file in xml format}
}
\value{
A data frame containing input data in long-format
}
\description{
Parse orthoXML input file
}
\examples{
\dontrun{
inputFile <- system.file(
    "extdata", "test.main.xml", package = "phyloprofile", mustWork = TRUE
)
xmlParser(inputFile)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
